requirements.txt
streamlit
tensorflow
numpy
pandas
opencv-python-headless
scikit-learn
matplotlib
seaborn
Pillow
